feat: add state to popup and update eventhandler
