+++ src/app.rs
@@ -1,4 +1,4 @@
-use crate::{events::EventHandler, screen::Screen};
+use crate::{events::EventHandler, popup::Popup, screen::Screen};
 use color_eyre::Result;
 use crossterm::{
     event::{DisableMouseCapture, EnableMouseCapture},
@@ -6,7 +6,11 @@ use crossterm::{
     ExecutableCommand,
 };
 use ratatui::{backend::CrosstermBackend, Terminal};
-use std::{cell::RefCell, io::{stdout, Stdout}, rc::Rc};
+use std::{
+    cell::RefCell,
+    io::{stdout, Stdout},
+    rc::Rc,
+};
 pub struct App {
     event_handler: EventHandler,
     screen: Rc<RefCell<Screen>>,
@@ -15,7 +19,8 @@ pub struct App {
 impl App {
     pub fn new() -> App {
         let screen = Rc::new(RefCell::new(Screen::new()));
-        let event_handler = EventHandler::new(Rc::clone(&screen));
+        let popup = Rc::new(RefCell::new(Popup::new()));
+        let event_handler = EventHandler::new(Rc::clone(&screen), Rc::clone(&popup));
         Self {
             event_handler,
             screen,
